+++ src/components/index.js
@@ -2,3 +2,4 @@ export { default as Datapairs } from "./Datapairs/Datapairs";
 export { default as BlockIndex } from "./BlockIndex/BlockIndex";
 export { default as Block } from "./Block/Block";
 export { default as HeaderCards } from "./HeaderCards/HeaderCards";
+export { default as ChainSelect } from "./ChainSelect/ChainSelect";