+++ src/components/ui/Heading/Heading.js
@@ -1,7 +1,29 @@
 import React from "react";
 
+import PropTypes from "prop-types";
+
+import { ChainSelect } from "@Components";
+
 import "./heading.scss";
 
-export default function Heading({ children }) {
-  return <h1 className="heading">{children}</h1>;
+const propTypes = {
+  primary: PropTypes.bool,
+};
+
+const defaultProps = {
+  primary: false,
 }
+
+function Heading({ primary, children }) {
+  return (
+    <h1 className="heading">
+      {children}
+      {primary && <ChainSelect />}
+    </h1>
+  );
+};
+
+Heading.propTypes = propTypes;
+Heading.defaultProps = defaultProps;
+
+export default Heading;