+++ src/utils/constants.js
@@ -3,6 +3,42 @@ export const tokenAuth = process.env.REACT_APP_TOKEN_AUTH;
 
 export const devEnvironments = ["dev", "sbx"];
 
-export const isDev = devEnvironments.some((environment) =>
+export const environment = devEnvironments.find((environment) =>
   process.env.REACT_APP_BLOCKS_API.includes(environment)
 );
+
+export const isDev = devEnvironments.includes(environment);
+
+export const availableChains = {
+  undefined: {
+    name: "BSC",
+    symbol: "bsc",
+    dev: "Testnet",
+    sbx: "Testnet",
+    prod: "Mainnet",
+  },
+  polygon: {
+    name: "Polygon",
+    symbol: "polygon",
+    dev: "Mumbai",
+    sbx: "Mumbai",
+    prod: "Mainnet",
+  },
+  ethereum: {
+    name: "Ethereum",
+    symbol: "eth",
+    dev: "Kovan",
+    sbx: "Ropsten",
+    prod: "Mainnet",
+  },
+  avax: {
+    name: "Avalanche",
+    symbol: "avax",
+    dev: "Testnet",
+    sbx: "Testnet",
+    prod: "Mainnet",
+  },
+};
+
+export const currentChain = availableChains[chainId].name;
+export const currentNetwork = availableChains[chainId][environment];