+++ CHANGELOG.md
@@ -5,11 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2021-11-18
+
+### Added
+
+- Chain selection
+- Search bar to block leaves keys
+- Border to select search
+- Add timestamp to age title
+- Add timestamp to block
+
 ## [1.1.3] - 2021-11-11
 
 ### Added
 
 - Makefile and github action to deploy to avalanche
+
 ## [1.1.2] - 2021-10-27
 
 ### Fix